+++ database/migrations/2024_09_25_065009_create_memberships_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->foreignId('user_id')->constrained('users')->onDelete('cascade');
             $table->integer('points')->default(0);
             $table->foreignId('rank_id')->constrained('membership_ranks');
-            $table->tinyInteger('status')->default(1)->comment('1:active; 2: inactive');
+            $table->integer('status')->default(1)->comment('1:active; 2: inactive');
             $table->integer('total_spent')->default(0);
             $table->timestamps();
         });
